+++ index.html
@@ -41,12 +41,20 @@
             <h3>Books You Are Looking For</h3>
             <div class="container__books-grid">         
             </div>
-  
+            <!-- <div id="myModal" class="modal"> -->
+
+                <!-- Modal content -->
+                <!-- <div class="modal-content"> -->
+                  <!-- <span class="close">&times;</span> -->
+                  <!-- <p>Some text in the Modal..</p> -->
+                <!-- </div> -->
+              
+              </div>
         </section>
     </main>
 <footer>
     <p class="text text--reg text--light-center"> &copy Copyright Martyna Krol <script>document.write(new Date().getFullYear());</script></p>
 </footer>
-    <script type = "module" src="./../main.js"></script>
+    <script type = "module" src="./main.js"></script>
 </body>
 </html>
