+++ modules/books-grid.scss
@@ -43,7 +43,7 @@ h4 {
 	flex-direction: column;
 	align-items: center;
 	text-align: left;
-	// cursor: pointer;
+	cursor: pointer;
 }
 
 .container__books-grid > div > * {
